+++ apps/web/CHANGELOG.md
@@ -1,5 +1,13 @@
 # web
 
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @astracms/core@1.0.2
+  - @astracms/astro-loader@1.0.2
+
 ## 0.0.2
 
 ### Patch Changes