+++ packages/astro-loader/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @astracms/astro-loader
 
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @astracms/core@1.0.2
+
 ## 1.0.0
 
 ### Major Changes