+++ packages/vue/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @astracms/vue
 
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @astracms/core@1.0.2
+
 ## 1.0.1
 
 ### Patch Changes